Fix build error on 64bit arch

+++ vcore/src/vcore/Certificate.cpp
@@ -138,7 +138,7 @@ bool Certificate::isSignedBy(const CertificatePtr &parent) const
 Certificate::Fingerprint Certificate::getFingerprint(
         Certificate::FingerprintType type) const
 {
-    size_t fingerprintlength = EVP_MAX_MD_SIZE;
+    unsigned int fingerprintlength = EVP_MAX_MD_SIZE;
     unsigned char fingerprint[EVP_MAX_MD_SIZE];
     Fingerprint raw;
